+++ yandex/resource_yandex_compute_instance_group.go
@@ -320,10 +320,6 @@ func resourceYandexComputeInstanceGroup() *schema.Resource {
 										Type:     schema.TypeInt,
 										Required: true,
 									},
-									"cpu_utilization_target": {
-										Type:     schema.TypeFloat,
-										Required: true,
-									},
 									"min_zone_size": {
 										Type:     schema.TypeInt,
 										Optional: true,
@@ -343,6 +339,36 @@ func resourceYandexComputeInstanceGroup() *schema.Resource {
 										Type:     schema.TypeInt,
 										Optional: true,
 										Computed: true,
+									},
+									"cpu_utilization_target": {
+										Type:     schema.TypeFloat,
+										Optional: true,
+									},
+									"custom_rule": {
+										Type:     schema.TypeList,
+										Optional: true,
+										Elem: &schema.Resource{
+											Schema: map[string]*schema.Schema{
+												"rule_type": {
+													Type:         schema.TypeString,
+													Required:     true,
+													ValidateFunc: validation.StringInSlice([]string{"UTILIZATION", "WORKLOAD"}, false),
+												},
+												"metric_type": {
+													Type:         schema.TypeString,
+													Required:     true,
+													ValidateFunc: validation.StringInSlice([]string{"GAUGE", "COUNTER"}, false),
+												},
+												"metric_name": {
+													Type:     schema.TypeString,
+													Required: true,
+												},
+												"target": {
+													Type:     schema.TypeFloat,
+													Required: true,
+												},
+											},
+										},
 									},
 								},
 							},